# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Home Lab Launcher API",
    "version": "0.7.0",
    "description": "Pre-1.0 JSON API for the bundled frontend, trusted plugins, and private-deployment automation. Mutating API routes require the session CSRF token in the X-CSRF-Token header, except /api/bootstrap and /api/auth/login. Automation may instead authenticate with an Admin-issued API token via Authorization: Bearer; token requests skip CSRF, act with the token's role, and are rejected on session/profile endpoints (/api/auth/*, /api/me*, /api/bootstrap*)."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
...
  "security": [
    { "sessionCookie": [] },
    { "bearerToken": [] }
  ],
  "paths": {
...
    "/settings": {
      "patch": {
        "tags": ["Settings"],
        "summary": "Update admin settings",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SettingsUpdateRequest" } } } },
        "responses": { "200": { "$ref": "#/components/responses/Ok" }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } }
      }
    },
...
    "/app-assets": {
      "post": {
        "tags": ["Assets", "Appearance"],
        "summary": "Upload or download a branding asset",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AppAssetRequest" } } } },
        "responses": { "201": { "description": "Created asset", "content": { "application/json": { "schema": { "type": "object", "required": ["url"], "properties": { "url": { "type": "string" } } } } } }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } }
      }
    },
...
    "/admin/api-tokens": {
      "get": { "tags": ["Admin"], "summary": "List API tokens", "description": "Metadata only; the token secret is never returned after creation.", "responses": { "200": { "description": "API tokens", "content": { "application/json": { "schema": { "type": "object", "required": ["tokens"], "properties": { "tokens": { "type": "array", "items": { "$ref": "#/components/schemas/ApiToken" } } } } } } }, "403": { "$ref": "#/components/responses/Error" } } },
      "post": { "tags": ["Admin"], "summary": "Create an API token", "description": "The plaintext token is returned exactly once and stored only as a SHA-256 hash.", "security": [{ "sessionCookie": [], "csrfToken": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiTokenCreateRequest" } } } }, "responses": { "201": { "description": "Created token", "content": { "application/json": { "schema": { "type": "object", "required": ["token", "apiToken"], "properties": { "token": { "type": "string", "description": "Shown-once plaintext bearer token (hll_ prefix)." }, "apiToken": { "$ref": "#/components/schemas/ApiToken" } } } } } }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/api-tokens/{id}": {
      "delete": { "tags": ["Admin"], "summary": "Revoke an API token", "security": [{ "sessionCookie": [], "csrfToken": [] }], "parameters": [{ "$ref": "#/components/parameters/Id" }], "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/overview": { "get": { "tags": ["Admin"], "summary": "Read admin overview", "responses": { "200": { "description": "Admin overview", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminOverview" } } } }, "403": { "$ref": "#/components/responses/Error" } } } },
...
    "/admin/config": { "get": { "tags": ["Admin"], "summary": "Read effective non-secret config", "responses": { "200": { "description": "Effective config", "content": { "application/json": { "schema": { "type": "object", "properties": { "config": { "$ref": "#/components/schemas/GenericObject" } } } } } }, "403": { "$ref": "#/components/responses/Error" } } } },
    "/admin/notices": { "get": { "tags": ["Admin"], "summary": "Read admin notices", "responses": { "200": { "description": "Notices", "content": { "application/json": { "schema": { "type": "object", "properties": { "notices": { "type": "array", "items": { "$ref": "#/components/schemas/Notice" } } } } } } }, "403": { "$ref": "#/components/responses/Error" } } } },
    "/admin/backup": { "get": { "tags": ["Admin"], "summary": "Export portable config backup", "responses": { "200": { "description": "Backup JSON", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Backup" } } } }, "403": { "$ref": "#/components/responses/Error" } } } },
    "/admin/database/export": { "get": { "tags": ["Admin"], "summary": "Download SQLite database export", "responses": { "200": { "description": "SQLite database file" }, "403": { "$ref": "#/components/responses/Error" } } } },
    "/admin/restore": {
      "post": {
        "tags": ["Admin"],
        "summary": "Restore supported settings and services from backup JSON",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "parameters": [{ "name": "preview", "in": "query", "schema": { "type": "boolean" }, "description": "When true, return preview only." }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Backup" } } } },
        "responses": { "200": { "$ref": "#/components/responses/Ok" }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/restore/preview": {
      "post": {
        "tags": ["Admin"],
        "summary": "Preview a config restore",
        "security": [{ "sessionCookie": [], "csrfToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Backup" } } } },
        "responses": { "200": { "description": "Restore preview", "content": { "application/json": { "schema": { "type": "object", "properties": { "preview": { "$ref": "#/components/schemas/GenericObject" } } } } } }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/logs": {
      "get": {
        "tags": ["Admin"],
        "summary": "Search audit logs",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 500 } },
          { "name": "level", "in": "query", "schema": { "type": "string" } },
          { "name": "action", "in": "query", "schema": { "type": "string" } },
          { "name": "actor", "in": "query", "schema": { "type": "string" } },
          { "name": "q", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Audit logs", "content": { "application/json": { "schema": { "type": "object", "properties": { "logs": { "type": "array", "items": { "$ref": "#/components/schemas/AuditLog" } } } } } } }, "403": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/logs/export": { "get": { "tags": ["Admin"], "summary": "Export audit logs", "responses": { "200": { "description": "Audit log export", "content": { "application/json": { "schema": { "type": "object", "properties": { "exportedAt": { "type": "string", "format": "date-time" }, "logs": { "type": "array", "items": { "$ref": "#/components/schemas/AuditLog" } } } } } } }, "403": { "$ref": "#/components/responses/Error" } } } },
    "/admin/logs/retention": { "patch": { "tags": ["Admin"], "summary": "Update audit-log retention", "security": [{ "sessionCookie": [], "csrfToken": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "days": { "type": "integer", "minimum": 1, "maximum": 3650 } } } } } }, "responses": { "200": { "description": "Retention", "content": { "application/json": { "schema": { "type": "object", "properties": { "days": { "type": "integer" } } } } } }, "403": { "$ref": "#/components/responses/Error" } } } },
    "/admin/logs/prune": { "post": { "tags": ["Admin"], "summary": "Prune old audit logs", "security": [{ "sessionCookie": [], "csrfToken": [] }], "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Error" } } } },
    "/admin/appearance": {
      "get": { "tags": ["Appearance"], "summary": "Read appearance settings", "responses": { "200": { "description": "Appearance", "content": { "application/json": { "schema": { "type": "object", "properties": { "appearance": { "$ref": "#/components/schemas/Appearance" }, "presets": { "type": "array", "items": { "$ref": "#/components/schemas/ThemePreset" } } } } } } }, "403": { "$ref": "#/components/responses/Error" } } },
      "put": { "tags": ["Appearance"], "summary": "Replace appearance settings", "security": [{ "sessionCookie": [], "csrfToken": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "appearance": { "$ref": "#/components/schemas/Appearance" } } } } } }, "responses": { "200": { "description": "Appearance", "content": { "application/json": { "schema": { "type": "object", "properties": { "appearance": { "$ref": "#/components/schemas/Appearance" } } } } } }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/appearance/reset": { "post": { "tags": ["Appearance"], "summary": "Reset appearance settings", "security": [{ "sessionCookie": [], "csrfToken": [] }], "responses": { "200": { "description": "Appearance", "content": { "application/json": { "schema": { "type": "object", "properties": { "appearance": { "$ref": "#/components/schemas/Appearance" } } } } } }, "403": { "$ref": "#/components/responses/Error" } } } },
    "/admin/theme-presets": {
      "get": { "tags": ["Appearance"], "summary": "List theme presets", "responses": { "200": { "description": "Theme presets", "content": { "application/json": { "schema": { "type": "object", "properties": { "presets": { "type": "array", "items": { "$ref": "#/components/schemas/ThemePreset" } } } } } } }, "403": { "$ref": "#/components/responses/Error" } } },
      "post": { "tags": ["Appearance"], "summary": "Create a theme preset", "security": [{ "sessionCookie": [], "csrfToken": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ThemePresetWriteRequest" } } } }, "responses": { "201": { "description": "Theme preset", "content": { "application/json": { "schema": { "type": "object", "properties": { "preset": { "$ref": "#/components/schemas/ThemePreset" } } } } } }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/theme-presets/{id}": {
      "patch": { "tags": ["Appearance"], "summary": "Update a theme preset", "security": [{ "sessionCookie": [], "csrfToken": [] }], "parameters": [{ "$ref": "#/components/parameters/Id" }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ThemePresetWriteRequest" } } } }, "responses": { "200": { "description": "Theme preset", "content": { "application/json": { "schema": { "type": "object", "properties": { "preset": { "$ref": "#/components/schemas/ThemePreset" } } } } } }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } },
      "delete": { "tags": ["Appearance"], "summary": "Delete a theme preset", "security": [{ "sessionCookie": [], "csrfToken": [] }], "parameters": [{ "$ref": "#/components/parameters/Id" }], "responses": { "200": { "$ref": "#/components/responses/Ok" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } }
    },
    "/admin/theme-presets/{id}/apply": { "post": { "tags": ["Appearance"], "summary": "Apply a theme preset", "security": [{ "sessionCookie": [], "csrfToken": [] }], "parameters": [{ "$ref": "#/components/parameters/Id" }], "responses": { "200": { "description": "Appearance", "content": { "application/json": { "schema": { "type": "object", "properties": { "appearance": { "$ref": "#/components/schemas/Appearance" } } } } } }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } } },
    "/admin/theme-presets/{id}/export": { "get": { "tags": ["Appearance"], "summary": "Export a theme preset", "parameters": [{ "$ref": "#/components/parameters/Id" }], "responses": { "200": { "description": "Theme preset export", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ThemePresetExport" } } } }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } } },
    "/admin/theme-presets/import": { "post": { "tags": ["Appearance"], "summary": "Import a theme preset", "security": [{ "sessionCookie": [], "csrfToken": [] }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ThemePresetExport" } } } }, "responses": { "201": { "description": "Theme preset", "content": { "application/json": { "schema": { "type": "object", "properties": { "preset": { "$ref": "#/components/schemas/ThemePreset" } } } } } }, "400": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } } } },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "hll.sid" },
      "csrfToken": { "type": "apiKey", "in": "header", "name": "X-CSRF-Token" },
      "bearerToken": { "type": "http", "scheme": "bearer", "description": "Admin-issued API token (hll_ prefix). Skips CSRF; not accepted on session/profile endpoints." }
    },
    "parameters": {
      "Id": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "responses": {
      "Ok": { "description": "Success", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Ok" } } } },
...
      "Error": { "description": "Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
    },
    "schemas": {
      "Ok": { "type": "object", "required": ["ok"], "properties": { "ok": { "type": "boolean" } }, "additionalProperties": true },
...
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "type": "string" } }, "additionalProperties": true },
      "GenericObject": { "type": "object", "additionalProperties": true },
      "Role": { "type": "string", "enum": ["admin", "editor", "user"] },
...
      "Appearance": { "type": "object", "properties": { "version": { "type": "integer" }, "brand": { "$ref": "#/components/schemas/GenericObject" }, "hero": { "$ref": "#/components/schemas/GenericObject" }, "theme": { "$ref": "#/components/schemas/GenericObject" } }, "additionalProperties": true },
      "PublicSettings": { "type": "object", "properties": { "appName": { "type": "string" }, "version": { "type": "string" }, "repositoryUrl": { "type": "string" }, "appBaseUrl": { "type": "string" }, "publicReadEnabled": { "type": "boolean" }, "appearance": { "$ref": "#/components/schemas/Appearance" } }, "additionalProperties": true },
      "SettingsUpdateRequest": { "type": "object", "properties": { "app_name": { "type": "string" }, "app_base_url": { "type": "string" }, "public_read_enabled": { "type": "boolean" }, "scheduled_backup_location": { "type": "string" }, "health_webhook_url": { "type": "string", "format": "uri", "description": "Webhook receiving a JSON POST on service up/down transitions. Empty string disables." }, "health_history_retention_days": { "type": "integer", "minimum": 1, "maximum": 90, "default": 7 } }, "additionalProperties": false },
...
      "AppAssetRequest": { "type": "object", "properties": { "assetData": { "type": "string", "description": "Data URL for uploaded image." }, "url": { "type": "string", "format": "uri", "description": "Remote image URL to download." }, "name": { "type": "string" } }, "additionalProperties": true },
...
      "Notice": { "type": "object", "properties": { "level": { "type": "string" }, "source": { "type": "string" }, "title": { "type": "string" }, "message": { "type": "string" }, "createdAt": { "type": "string" } }, "additionalProperties": true },
      "AdminOverview": { "type": "object", "properties": { "settings": { "$ref": "#/components/schemas/PublicSettings" }, "counts": { "$ref": "#/components/schemas/GenericObject" }, "runtime": { "$ref": "#/components/schemas/GenericObject" }, "warnings": { "type": "array", "items": { "$ref": "#/components/schemas/Notice" } }, "notices": { "type": "array", "items": { "$ref": "#/components/schemas/Notice" } } }, "additionalProperties": true },
      "Backup": { "type": "object", "properties": { "exportedAt": { "type": "string" }, "format": { "type": "string", "const": "home-lab-launcher-config-v1" }, "appVersion": { "type": "string" }, "settings": { "$ref": "#/components/schemas/GenericObject" }, "services": { "type": "array", "items": { "$ref": "#/components/schemas/Service" } }, "users": { "type": "array", "items": { "$ref": "#/components/schemas/User" } }, "plugins": { "type": "array", "items": { "$ref": "#/components/schemas/Plugin" } } }, "additionalProperties": true },
      "AuditLog": { "type": "object", "properties": { "id": { "type": "integer" }, "level": { "type": "string" }, "action": { "type": "string" }, "actorUserId": { "type": ["integer", "null"] }, "actorUsername": { "type": ["string", "null"] }, "ip": { "type": ["string", "null"] }, "details": { "$ref": "#/components/schemas/GenericObject" }, "createdAt": { "type": "string" } }, "additionalProperties": true },
      "ThemePreset": { "type": "object", "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "description": { "type": "string" }, "appearance": { "$ref": "#/components/schemas/Appearance" }, "createdAt": { "type": "string" }, "updatedAt": { "type": "string" } }, "additionalProperties": true },
      "ThemePresetWriteRequest": { "type": "object", "properties": { "id": { "type": "string" }, "name": { "type": "string" }, "description": { "type": "string" }, "appearance": { "$ref": "#/components/schemas/Appearance" } }, "additionalProperties": true },
      "ThemePresetExport": { "type": "object", "required": ["format", "name", "appearance"], "properties": { "format": { "type": "string", "const": "home-lab-launcher-theme-v1" }, "name": { "type": "string" }, "description": { "type": "string" }, "appearance": { "$ref": "#/components/schemas/Appearance" } }, "additionalProperties": true },
...
      "ApiToken": { "type": "object", "properties": { "id": { "type": "integer" }, "name": { "type": "string" }, "tokenPrefix": { "type": "string", "description": "Display prefix of the token for identification; the full secret is never stored." }, "role": { "$ref": "#/components/schemas/Role" }, "createdBy": { "type": ["integer", "null"] }, "expiresAt": { "type": ["string", "null"] }, "lastUsedAt": { "type": ["string", "null"] }, "createdAt": { "type": "string" } }, "additionalProperties": true },
      "ApiTokenCreateRequest": { "type": "object", "required": ["name"], "properties": { "name": { "type": "string", "maxLength": 80 }, "role": { "$ref": "#/components/schemas/Role" }, "expiresDays": { "type": "integer", "minimum": 1, "maximum": 3650, "description": "Optional expiry in days from creation." } }, "additionalProperties": false }
    }
  }
}